don't show save when hosted
