+++ package.json
@@ -100,11 +100,11 @@
         }
     },
     "devDependencies": {
-        "@inquirer/prompts": "^7.3.1",
+        "@inquirer/prompts": "^7.3.2",
         "glob": "^11.0.1",
         "npm-check-updates": "^17.1.14",
         "npm-run-all": "^4.1.5",
-        "prettier": "^3.5.0",
+        "prettier": "^3.5.1",
         "zx": "^8.3.2"
     }
 }
