+++ packages/cli/package.json
@@ -53,8 +53,8 @@
   "dependencies": {
     "@anthropic-ai/bedrock-sdk": "^0.12.4",
     "@anthropic-ai/sdk": "^0.36.3",
-    "@azure/identity": "^4.6.0",
-    "@inquirer/prompts": "^7.3.1",
+    "@azure/identity": "^4.7.0",
+    "@inquirer/prompts": "^7.3.2",
     "@modelcontextprotocol/sdk": "^1.5.0",
     "@octokit/plugin-paginate-rest": "^11.4.0",
     "@octokit/plugin-retry": "^7.1.3",
@@ -70,7 +70,7 @@
     "mathjs": "^14.2.1",
     "pyodide": "^0.27.2",
     "tabletojson": "^4.1.5",
-    "tsx": "^4.19.2",
+    "tsx": "^4.19.3",
     "turndown": "^7.2.0",
     "turndown-plugin-gfm": "^1.0.2",
     "typescript": "5.7.3",
@@ -99,7 +99,7 @@
     "@types/dockerode": "^3.3.34",
     "@types/fs-extra": "^11.0.4",
     "@types/memorystream": "^0.3.4",
-    "@types/node": "^22.13.1",
+    "@types/node": "^22.13.4",
     "@types/papaparse": "^5.3.15",
     "@types/prompts": "^2.4.9",
     "@types/replace-ext": "^2.0.2",
@@ -117,7 +117,7 @@
     "memorystream": "^0.3.1",
     "node-sarif-builder": "^3.2.0",
     "octokit": "^4.1.1",
-    "openai": "^4.84.0",
+    "openai": "^4.85.2",
     "pretty-bytes": "^6.1.1",
     "replace-ext": "^2.0.0",
     "ws": "^8.18.0",
