+++ packages/core/package.json
@@ -20,7 +20,7 @@
   "devDependencies": {
     "@anthropic-ai/bedrock-sdk": "^0.12.4",
     "@anthropic-ai/sdk": "^0.36.3",
-    "@azure/identity": "^4.6.0",
+    "@azure/identity": "^4.7.0",
     "@elastic/micro-jq": "^1.8.0",
     "@huggingface/jinja": "^0.3.3",
     "@huggingface/transformers": "^3.3.3",
@@ -38,7 +38,7 @@
     "@types/ini": "^4.1.1",
     "@types/mime-types": "^2.1.4",
     "@types/mustache": "^4.2.5",
-    "@types/node": "^22.13.1",
+    "@types/node": "^22.13.4",
     "@types/object-inspect": "^1.13.0",
     "@types/semver": "^7.5.8",
     "@types/shell-quote": "^1.7.5",
@@ -73,14 +73,14 @@
     "merge-descriptors": "^2.0.0",
     "mime": "^4.0.6",
     "minimatch": "^10.0.1",
-    "minisearch": "^7.1.1",
+    "minisearch": "^7.1.2",
     "mustache": "^4.2.0",
     "object-inspect": "^1.13.4",
-    "openai": "^4.84.0",
+    "openai": "^4.85.2",
     "p-limit": "^6.2.0",
     "package-manager-detector": "^0.2.9",
     "parse-diff": "^0.11.1",
-    "prettier": "^3.5.0",
+    "prettier": "^3.5.1",
     "pretty-bytes": "^6.1.1",
     "sanitize-html": "^2.14.0",
     "semver": "^7.7.1",
@@ -90,7 +90,7 @@
     "toml": "^3.0.0",
     "tree-sitter-wasms": "^0.1.11",
     "ts-dedent": "^2.2.0",
-    "tsx": "^4.19.2",
+    "tsx": "^4.19.3",
     "typescript": "5.7.3",
     "vectra": "^0.9.0",
     "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.2/xlsx-0.20.2.tgz",
