+++ packages/sample/package.json
@@ -15,10 +15,10 @@
     "test:scripts:view": "node ../cli/built/genaiscript.cjs test view"
   },
   "devDependencies": {
-    "@agentic/calculator": "^7.2.0",
-    "@agentic/core": "^7.2.0",
-    "@agentic/weather": "^7.2.0",
-    "@azure/identity": "^4.6.0",
+    "@agentic/calculator": "^7.3.1",
+    "@agentic/core": "^7.3.1",
+    "@agentic/weather": "^7.3.1",
+    "@azure/identity": "^4.7.0",
     "@azure/storage-blob": "^12.26.0",
     "@huggingface/transformers": "^3.3.3",
     "@modelcontextprotocol/sdk": "^1.5.0",
