+++ packages/web/package.json
@@ -12,9 +12,9 @@
   "devDependencies": {
     "@daiji256/rehype-mathml": "^1.1.1",
     "@types/marked": "^6.0.0",
-    "@types/react": "^19.0.8",
-    "@types/react-dom": "^19.0.3",
-    "@vscode-elements/elements": "1.12.0",
+    "@types/react": "^19.0.10",
+    "@types/react-dom": "^19.0.4",
+    "@vscode-elements/elements": "1.13.1",
     "@vscode-elements/webview-playground": "^1.6.0",
     "@vscode/codicons": "^0.0.36",
     "clsx": "^2.1.1",
